refactor: clean-up part of obsolete docs and scripts
