+++ package.json
@@ -12,7 +12,6 @@
   ],
   "scripts": {
     "postinstall": "husky install",
-    "build:apps": "yarn compile && hardhat run --no-compile scripts/build-apps-frontend.js",
     "lint": "yarn lint:sol && yarn lint:js",
     "lint:sol": "solhint \"contracts/**/*.sol\" --ignore-path .solhintignore",
     "lint:sol:fix": "yarn lint:sol --fix",
@@ -35,28 +34,8 @@
     "extract-artifacts": "node ./scripts/extract-artifacts.js",
     "size-contracts": "yarn run hardhat size-contracts",
     "deploy": "yarn run deploy:all",
-    "deploy:all": "yarn compile && yarn deploy:aragon-env && yarn deploy:aragon-std-apps && yarn deploy:apm-and-template && yarn deploy:apps && yarn deploy:dao",
-    "deploy:e2e:all": "yarn compile && yarn deploy:e2e:aragon-env && yarn deploy:e2e:aragon-std-apps && yarn deploy:e2e:apm-and-template && yarn deploy:e2e:apps && yarn deploy:e2e:dao",
-    "deploy:aragon-env": "hardhat --network $NETWORK run --no-compile ./scripts/scratch/deploy-aragon-env.js",
-    "deploy:e2e:aragon-env": "NETWORK_NAME=e2e NETWORK_STATE_FILE=deployed-e2e.json hardhat run --no-compile ./scripts/scratch/deploy-aragon-env.js",
-    "deploy:aragon-std-apps": "RELEASE_TYPE=major hardhat run --no-compile ./scripts/scratch/deploy-aragon-std-apps.js --config ./hardhat.config.aragon-apps.js",
-    "deploy:e2e:aragon-std-apps": "NETWORK_NAME=e2e NETWORK_STATE_FILE=deployed-e2e.json IPFS_API_URL=http://localhost:5001/api/v0 IPFS_GATEWAY_URL=http://localhost:8080 RELEASE_TYPE=major hardhat run --no-compile ./scripts/scratch/deploy-aragon-std-apps.js --config ./hardhat.config.aragon-apps.js",
-    "deploy:apm-and-template": "hardhat run --no-compile  ./scripts/scratch/deploy-lido-apm-and-template.js",
-    "deploy:e2e:apm-and-template": "NETWORK_NAME=e2e NETWORK_STATE_FILE=deployed-e2e.json hardhat run --no-compile  ./scripts/scratch/deploy-lido-apm-and-template.js",
-    "deploy:apps": "hardhat run --no-compile ./scripts/scratch/deploy-lido-apps.js",
-    "deploy:e2e:apps": "NETWORK_NAME=e2e NETWORK_STATE_FILE=deployed-e2e.json IPFS_API_URL=http://localhost:5001/api/v0 IPFS_GATEWAY_URL=http://localhost:8080 hardhat run --no-compile ./scripts/scratch/deploy-lido-apps.js",
     "deploy:app-pool": "APPS=depool yarn deploy:apps",
     "deploy:app-staking-providers-registry": "APPS=staking-providers-registry yarn deploy:apps",
-    "deploy:local:all": "yarn compile && yarn deploy:local:aragon-env && yarn deploy:local:aragon-std-apps && yarn deploy:local:apm-and-template && yarn deploy:local:apps && yarn deploy:local:dao",
-    "deploy:local:aragon-env": "NETWORK_NAME=local NETWORK_STATE_FILE=deployed-local.json hardhat run --no-compile ./scripts/scratch/deploy-aragon-env.js",
-    "deploy:local:aragon-std-apps": "NETWORK_NAME=local NETWORK_STATE_FILE=deployed-local.json IPFS_API_URL=http://127.0.0.1:5001/api/v0 IPFS_GATEWAY_URL=http://127.0.0.1:8080 RELEASE_TYPE=major hardhat run --no-compile ./scripts/scratch/deploy-aragon-std-apps.js --config ./hardhat.config.aragon-apps.js",
-    "deploy:rinkeby:aragon-env": "NETWORK_NAME=rinkeby NETWORK_STATE_FILE=deployed-rinkeby.json hardhat run --no-compile ./scripts/scratch/deploy-aragon-env.js",
-    "deploy:rinkeby:aragon-std-apps": "NETWORK_NAME=rinkeby NETWORK_STATE_FILE=deployed-rinkeby.json IPFS_API_URL=http://127.0.0.1:5001/api/v0 IPFS_GATEWAY_URL=http://127.0.0.1:8080 RELEASE_TYPE=major hardhat run --no-compile ./scripts/scratch/deploy-aragon-std-apps.js --config ./hardhat.config.aragon-apps.js",
-    "deploy:mainnetfork:aragon-env": "NETWORK_NAME=mainnetfork NETWORK_STATE_FILE=deployed-mainnetfork.json hardhat run --no-compile ./scripts/scratch/deploy-aragon-env.js",
-    "deploy:mainnetfork:aragon-std-apps": "NETWORK_NAME=mainnetfork NETWORK_STATE_FILE=deployed-mainnetfork.json IPFS_API_URL=http://127.0.0.1:5001/api/v0 IPFS_GATEWAY_URL=http://127.0.0.1:8080 RELEASE_TYPE=major hardhat run --no-compile ./scripts/scratch/deploy-aragon-std-apps.js --config ./hardhat.config.aragon-apps.js",
-    "deploy:aragon-std-apps-custom": "hardhat run  --no-compile ./scripts/scratch/custom-deploy-aragon-std-apps.js  --network $NETWORK --config ./hardhat.config.aragon-apps.js",
-    "deploy:local:apm-and-template": "NETWORK_NAME=local NETWORK_STATE_FILE=deployed-local.json hardhat run --no-compile  ./scripts/scratch/deploy-lido-apm-and-template.js",
-    "deploy:local:apps": "NETWORK_NAME=local NETWORK_STATE_FILE=deployed-local.json IPFS_API_URL=http://127.0.0.1:5001/api/v0 IPFS_GATEWAY_URL=http://127.0.0.1:8080o hardhat run --no-compile ./scripts/scratch/deploy-lido-apps.js",
     "apps:lido": "yarn --cwd ./apps/lido/app/ cli",
     "apps:nos": "yarn --cwd ./apps/node-operators-registry/app/ cli",
     "lido:apps": "concurrently \"yarn apps:lido\" \"yarn apps:oracle\" \"yarn apps:nos\" ",
